[refactor] 패키지 수정 및 FIXME 주석 추가
